BetterDisplay support
If the audio device fails to set volume or mute:

- Check if the device is a possible display (connected via HDMI or DisplayPort)
- Check if BetterDisplay app is installed
- Execute the desired volume/mute option in BetterDisplay

Why:

Some display audio devices (if not all) can't be controlled directly with macOS volume options or MultiSoundChanger. BetterDisplay makes this possible.

When controlling the volume of a multi-output device BetterDisplay magic is not applied, so this implementation solves this scenario.

https://github.com/waydabber/BetterDisplay/discussions/2194